# MortgageCalculator/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Mortgage Calculator API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/payment-schedule/{mortgage_program_id}": {
            "post": {
                "tags": [
                    "/payment-schedule"
                ],
                "operationId": "7621eb27683cdac41b017c24a009a043",
                "parameters": [
                    {
                        "name": "mortgage_program_id",
                        "in": "path",
                        "description": "Buscar por estado",
                        "required": true
                    }
                ],
                "requestBody": {
                    "description": "Pass user credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "age",
                                    "is_a_bank_client",
                                    "property_value",
                                    "initial_payment",
                                    "loan_term",
                                    "property_type"
                                ],
                                "properties": {
                                    "property_value": {
                                        "type": "decimal",
                                        "format": "property_value",
                                        "example": "245000"
                                    },
                                    "initial_payment": {
                                        "type": "decimal",
                                        "format": "initial_payment",
                                        "example": "125000"
                                    },
                                    "loan_term": {
                                        "type": "integer",
                                        "format": "loan_term",
                                        "example": "24"
                                    },
                                    "payment_type": {
                                        "type": "string",
                                        "format": "property_type",
                                        "example": "аннуитетный"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "422": {
                        "description": "Wrong credentials response"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/users/{userId}": {
            "get": {
                "tags": [
                    "/users"
                ],
                "operationId": "cef8a64dad4500c2b80708ac3f63c2b2",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "path",
                        "description": "Buscar por estado",
                        "required": true
                    }
                ],
                "responses": {
                    "422": {
                        "description": "Wrong credentials response"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/users/{userId}/change_role": {
            "post": {
                "tags": [
                    "/users"
                ],
                "operationId": "3360eceecdfe28dc0e0b4b58d77cb3b6",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "path",
                        "description": "Buscar por estado",
                        "required": true
                    }
                ],
                "requestBody": {
                    "description": "Pass user credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "role": {
                                        "type": "string",
                                        "format": "role",
                                        "example": "moderator"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "422": {
                        "description": "Wrong credentials response"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}